refactor(rules): split the long rule files; extract shared helpers
Five rules grew past 250 lines with internally separable concerns:
`macro_trailing_comma`, `unicode_ellipsis_in_panic_messages`,
`prefer_raw_string`, `derive_ordering`, and
`single_letter_closure_param`. Each becomes a directory beside its
flat `<rule>.rs` entry, mirroring the existing `macro_argument_binding`
layout (the flat pattern is required by the rule's own
`flat_module_pattern` lint).

Two helpers shared across rules were already flagged in code
comments and are now factored out into dedicated crate-internal
modules:

- `src/enclosing_hir.rs` — the deepest-enclosing-HIR-node walker
  that `macro_trailing_comma` and `macro_argument_binding` both use
  for their pre-expansion -> late pass split. The duplication was
  called out by name in `macro_argument_binding/late.rs`.
- `src/literal_scan.rs` — the flagged-character emit loop and the
  `string_literal_quote_lengths` parser shared between the two
  Unicode-ellipsis rules.

Trivial cross-rule helpers land in the existing flat modules:
`parse_path_list` and `merge_with_builtins` move into `macro_path`,
where `parse_path` already lives; `binding_hir_id` moves into
`common.rs` next to its `binding_ident` sibling.

No lint behaviour changes; the diffs are purely structural.
