+++ src/rules/single_letter_closure_param.rs
@@ -1,13 +1,23 @@
+//! `perfectionist::single_letter_closure_param` — flag closure
+//! parameters whose identifier is one ASCII letter, unless the
+//! closure is a trivial single-expression callback.
+//!
+//! The trivial-callback predicate lives in [`triviality`]; this file
+//! owns the lint declaration, the configuration, and the late pass
+//! that drives them.
+
 use std::collections::BTreeSet;
 
 use clippy_utils::diagnostics::span_lint_and_help;
 use rustc_hir as hir;
-use rustc_hir::def::Res;
 use rustc_lint::{LateContext, LateLintPass, LintStore};
 use rustc_session::{declare_tool_lint, impl_lint_pass};
-use rustc_span::Symbol;
 
 use crate::common::{binding_ident, is_single_ascii_letter};
+
+mod triviality;
+
+use triviality::{is_trivial_wrapper, parent_call_callee_name, single_expression_body};
 
 declare_tool_lint! {
     /// ### What it does
@@ -190,93 +200,9 @@ impl SingleLetterClosureParam {
         lint_context: &LateContext<'tcx>,
         closure_expr: &'tcx hir::Expr<'tcx>,
     ) -> bool {
-        let parent = lint_context.tcx.parent_hir_node(closure_expr.hir_id);
-        let hir::Node::Expr(parent_expr) = parent else {
-            return false;
-        };
-        let callee_name = match parent_expr.kind {
-            hir::ExprKind::MethodCall(segment, _, _, _) => Some(segment.ident.name),
-            hir::ExprKind::Call(callee, _) => path_final_segment(callee),
-            _ => None,
-        };
-        let Some(name) = callee_name else {
+        let Some(name) = parent_call_callee_name(lint_context, closure_expr) else {
             return false;
         };
         self.comparison_methods.contains(name.as_str())
-    }
-}
-
-fn binding_hir_id<'hir>(pat: &'hir hir::Pat<'hir>) -> Option<hir::HirId> {
-    match pat.kind {
-        hir::PatKind::Binding(_, hir_id, _, None) => Some(hir_id),
-        _ => None,
-    }
-}
-
-/// If `body.value` is a single expression — either directly or
-/// wrapped in a block with no statements — return that
-/// expression. Otherwise return `None`.
-fn single_expression_body<'hir>(body: &'hir hir::Body<'hir>) -> Option<&'hir hir::Expr<'hir>> {
-    let value = body.value;
-    if let hir::ExprKind::Block(block, _) = value.kind {
-        if !block.stmts.is_empty() {
-            return None;
-        }
-        block.expr
-    } else {
-        Some(value)
-    }
-}
-
-fn path_final_segment<'hir>(expr: &'hir hir::Expr<'hir>) -> Option<Symbol> {
-    let hir::ExprKind::Path(qpath) = &expr.kind else {
-        return None;
-    };
-    let segment = match qpath {
-        hir::QPath::Resolved(_, path) => path.segments.last()?,
-        hir::QPath::TypeRelative(_, segment) => *segment,
-    };
-    Some(segment.ident.name)
-}
-
-/// Returns whether `expr` is a "trivial wrapper" around one of
-/// the closure's parameters:
-/// - a field access `param.field`,
-/// - a method call `param.foo(args)`,
-/// - a one-argument call `f(param)`,
-/// - a reference `&param`.
-///
-/// In every shape, `*` / `&` operators around the reference to
-/// the parameter are peeled by `is_param_ref` before matching,
-/// so `|s| (*s).foo()` and `|s| f(&*s)` both qualify.
-fn is_trivial_wrapper<'hir>(expr: &'hir hir::Expr<'hir>, params: &'hir [hir::Param<'hir>]) -> bool {
-    /// "Refers to a parameter, possibly through one or more `*` /
-    /// `&` operators." Peeling through these keeps `|s| (*s).foo()`
-    /// classified as a trivial wrapper, since the deref is a
-    /// purely-structural step the reader does not need help with.
-    fn is_param_ref(expr: &hir::Expr<'_>, params: &[hir::Param<'_>]) -> bool {
-        let mut expr = expr;
-        loop {
-            match &expr.kind {
-                hir::ExprKind::Unary(hir::UnOp::Deref, inner)
-                | hir::ExprKind::AddrOf(_, _, inner) => expr = inner,
-                hir::ExprKind::Path(hir::QPath::Resolved(None, path)) => {
-                    let Res::Local(local_hir_id) = path.res else {
-                        return false;
-                    };
-                    return params
-                        .iter()
-                        .any(|param| binding_hir_id(param.pat) == Some(local_hir_id));
-                }
-                _ => return false,
-            }
-        }
-    }
-    match expr.kind {
-        hir::ExprKind::Field(receiver, _) => is_param_ref(receiver, params),
-        hir::ExprKind::MethodCall(_, receiver, _, _) => is_param_ref(receiver, params),
-        hir::ExprKind::Call(_, args) => args.len() == 1 && is_param_ref(&args[0], params),
-        hir::ExprKind::AddrOf(_, _, inner) => is_param_ref(inner, params),
-        _ => false,
     }
 }
